pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rpg-project"
version = "0.1.0"
description = "A 2D RPG Game Server and Client"
authors = ["GitHub Copilot <copilot@github.com>"]
readme = "README.md"
packages = [
    { include = "server" },
    { include = "client" },
    { include = "common" },
]

[tool.poetry.dependencies]
python = "~3.12"
pydantic = "2.11.9"
msgpack = "1.1.1"
websockets = "15.0.1"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
fastapi = "0.116.1"
uvicorn = { version = "0.35.0", extras = ["standard"] }
sqlalchemy = "2.0.43"
alembic = "1.16.5"
asyncpg = "0.30.0"
psycopg2-binary = "2.9.9"
valkey-glide = "1.3.0"
pydantic-settings = "2.10.1"
passlib = "^1.7.4"
python-jose = { version = "3.5.0", extras = ["cryptography"] }
python-multipart = "0.0.20"
pyyaml = "^6.0"
pytmx = "^3.32"
pygame = "2.6.0"
prometheus-client = "^0.21.1"
python-json-logger = "^2.0.7"

[tool.poetry.group.client]
optional = true

[tool.poetry.group.client.dependencies]
pygame = "2.6.0"
aiohttp = "3.10.5"
pydantic = "2.11.9"
msgpack = "1.1.1"
websockets = "15.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2"
pytest-asyncio = ">=1.2.0"
pytest-xdist = ">=3.8.0"
httpx = "0.28.1"
Faker = ">=37.8.0"
aiosqlite = "^0.19.0"
